+++ src/python/pants/backend/codegen/protobuf/python/rules_integration_test.py
@@ -10,7 +10,7 @@ from pants.backend.codegen.protobuf.python import additional_fields
 from pants.backend.codegen.protobuf.python.rules import GeneratePythonFromProtobufRequest
 from pants.backend.codegen.protobuf.python.rules import rules as protobuf_rules
 from pants.backend.codegen.protobuf.target_types import ProtobufLibrary, ProtobufSources
-from pants.core.util_rules import determine_source_files, strip_source_roots
+from pants.core.util_rules import source_files, stripped_source_files
 from pants.engine.addresses import Address
 from pants.engine.internals.scheduler import ExecutionError
 from pants.engine.rules import RootRule
@@ -37,8 +37,8 @@ class ProtobufPythonIntegrationTest(ExternalToolTestBase):
             *super().rules(),
             *protobuf_rules(),
             *additional_fields.rules(),
-            *determine_source_files.rules(),
-            *strip_source_roots.rules(),
+            *source_files.rules(),
+            *stripped_source_files.rules(),
             RootRule(GeneratePythonFromProtobufRequest),
         )
