+++ src/python/pants/backend/python/dependency_inference/module_mapper_test.py
@@ -18,7 +18,7 @@ from pants.backend.python.dependency_inference.module_mapper import (
 )
 from pants.backend.python.target_types import PythonLibrary, PythonRequirementLibrary
 from pants.build_graph.build_file_aliases import BuildFileAliases
-from pants.core.util_rules import determine_source_files, strip_source_roots
+from pants.core.util_rules import source_files, stripped_source_files
 from pants.engine.addresses import Address
 from pants.engine.rules import RootRule
 from pants.python.python_requirement import PythonRequirement
@@ -83,8 +83,8 @@ class ModuleMapperTest(TestBase):
     def rules(cls):
         return (
             *super().rules(),
-            *strip_source_roots.rules(),
-            *determine_source_files.rules(),
+            *stripped_source_files.rules(),
+            *source_files.rules(),
             map_first_party_modules_to_addresses,
             map_module_to_address,
             map_third_party_modules_to_addresses,
