+++ src/python/pants/backend/python/lint/bandit/rules.py
@@ -21,8 +21,8 @@ from pants.core.goals.lint import (
     LintResults,
     LintSubsystem,
 )
-from pants.core.util_rules import determine_source_files, strip_source_roots
-from pants.core.util_rules.determine_source_files import SourceFiles, SourceFilesRequest
+from pants.core.util_rules import source_files, stripped_source_files
+from pants.core.util_rules.source_files import SourceFiles, SourceFilesRequest
 from pants.engine.fs import (
     Digest,
     DigestSubset,
@@ -173,7 +173,7 @@ def rules():
     return [
         *collect_rules(),
         UnionRule(LintRequest, BanditRequest),
-        *determine_source_files.rules(),
         *pex.rules(),
-        *strip_source_roots.rules(),
+        *source_files.rules(),
+        *stripped_source_files.rules(),
     ]
