+++ src/python/pants/backend/python/lint/docformatter/rules.py
@@ -17,8 +17,8 @@ from pants.backend.python.rules.pex import (
 from pants.backend.python.target_types import PythonSources
 from pants.core.goals.fmt import FmtResult
 from pants.core.goals.lint import LintRequest, LintResult, LintResults
-from pants.core.util_rules import determine_source_files, strip_source_roots
-from pants.core.util_rules.determine_source_files import SourceFiles, SourceFilesRequest
+from pants.core.util_rules import source_files, stripped_source_files
+from pants.core.util_rules.source_files import SourceFiles, SourceFilesRequest
 from pants.engine.fs import Digest, MergeDigests
 from pants.engine.process import FallibleProcessResult, Process, ProcessResult
 from pants.engine.rules import Get, MultiGet, collect_rules, rule
@@ -138,7 +138,7 @@ def rules():
         *collect_rules(),
         UnionRule(PythonFmtRequest, DocformatterRequest),
         UnionRule(LintRequest, DocformatterRequest),
-        *determine_source_files.rules(),
         *pex.rules(),
-        *strip_source_roots.rules(),
+        *source_files.rules(),
+        *stripped_source_files.rules(),
     ]
