+++ src/python/pants/backend/python/lint/python_fmt.py
@@ -7,7 +7,7 @@ from typing import Iterable, List, Type
 from pants.backend.python.target_types import PythonSources
 from pants.core.goals.fmt import FmtResult, LanguageFmtResults, LanguageFmtTargets
 from pants.core.goals.style_request import StyleRequest
-from pants.core.util_rules.determine_source_files import SourceFiles, SourceFilesRequest
+from pants.core.util_rules.source_files import SourceFiles, SourceFilesRequest
 from pants.engine.fs import Digest, Snapshot
 from pants.engine.rules import Get, collect_rules, rule
 from pants.engine.unions import UnionMembership, UnionRule, union
