+++ src/python/pants/backend/python/rules/run_setup_py_test.py
@@ -38,8 +38,7 @@ from pants.backend.python.target_types import (
 )
 from pants.build_graph.build_file_aliases import BuildFileAliases
 from pants.core.target_types import Files, Resources
-from pants.core.util_rules.determine_source_files import rules as determine_source_files_rules
-from pants.core.util_rules.strip_source_roots import rules as strip_source_roots_rules
+from pants.core.util_rules import source_files, stripped_source_files
 from pants.engine.addresses import Address
 from pants.engine.fs import Snapshot
 from pants.engine.internals.scheduler import ExecutionError
@@ -239,8 +238,8 @@ class TestGetSources(TestSetupPyBase):
             get_sources,
             RootRule(SetupPySourcesRequest),
             RootRule(SourceRootConfig),
-            *determine_source_files_rules(),
-            *strip_source_roots_rules(),
+            *source_files.rules(),
+            *stripped_source_files.rules(),
             *python_sources.rules(),
         ]
