+++ src/python/pants/backend/python/typecheck/mypy/rules.py
@@ -17,7 +17,7 @@ from pants.backend.python.rules.python_sources import rules as python_sources_ru
 from pants.backend.python.target_types import PythonSources
 from pants.backend.python.typecheck.mypy.subsystem import MyPy
 from pants.core.goals.typecheck import TypecheckRequest, TypecheckResult, TypecheckResults
-from pants.core.util_rules import determine_source_files, pants_bin, strip_source_roots
+from pants.core.util_rules import pants_bin, source_files, stripped_source_files
 from pants.engine.addresses import Addresses
 from pants.engine.fs import (
     CreateDigest,
@@ -124,9 +124,9 @@ def rules():
     return [
         *collect_rules(),
         UnionRule(TypecheckRequest, MyPyRequest),
-        *determine_source_files.rules(),
+        *source_files.rules(),
         *pants_bin.rules(),
         *pex_rules(),
         *python_sources_rules(),
-        *strip_source_roots.rules(),
+        *stripped_source_files.rules(),
     ]
