+++ src/python/pants/core/register.py
@@ -10,12 +10,12 @@ from pants.core.goals import binary, fmt, lint, repl, run, test, typecheck
 from pants.core.target_types import Files, GenericTarget, Resources
 from pants.core.util_rules import (
     archive,
-    determine_source_files,
     distdir,
     external_tool,
     filter_empty_sources,
     pants_bin,
-    strip_source_roots,
+    source_files,
+    stripped_source_files,
 )
 from pants.source import source_root
 
@@ -31,11 +31,11 @@ def rules():
         *test.rules(),
         *typecheck.rules(),
         # util_rules
-        *determine_source_files.rules(),
         *distdir.rules(),
         *filter_empty_sources.rules(),
         *pants_bin.rules(),
-        *strip_source_roots.rules(),
+        *source_files.rules(),
+        *stripped_source_files.rules(),
         *archive.rules(),
         *external_tool.rules(),
         *source_root.rules(),
